% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCapHist.R
\name{buildCapHist}
\alias{buildCapHist}
\title{Build capture history}
\usage{
buildCapHist(filter_ch = NULL, parent_child = NULL, configuration = NULL, ...)
}
\arguments{
\item{filter_ch}{filtered capture history as returned by the `filterDetections()` function, which has then been verified by a user and all blank or NA `user_keep_obs` entries have been completed.}

\item{...}{other inputs to the `defineCapHistCols` if desired.}
}
\description{
Based on a compressed and filtered capture history, translate into a capture history matrix.
}
\examples{
buildConfig()

}
\author{
Kevin See
}
