% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNodeGraph.R
\name{buildNodeGraph}
\alias{buildNodeGraph}
\title{Build Graph of Nodes}
\usage{
buildNodeGraph(parent_child = NULL)
}
\arguments{
\item{parent_child}{The output of `buildParentChild()`.}
}
\value{
A `tbl_graph` or `igraph` object
}
\description{
Create a graph object consisting of detection nodes and the edges that connect them.
}
\author{
Kevin See
}
