% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildParentChild.R
\name{buildParentChild}
\alias{buildParentChild}
\title{Build a Parent-Child table}
\usage{
buildParentChild(
  sites_sf = NULL,
  flowlines = NULL,
  rm_na_parent = F,
  add_rkm = F
)
}
\arguments{
\item{sites_sf}{an `sf` object containing the `SiteID` and location of all detection sites.}

\item{flowlines}{output from `queryFlowlines()` function.}

\item{rm_na_parent}{should rows with NA as the parent be automatically removed? Default is `FALSE`.}

\item{add_rkm}{should the RKM of the parent and child be added, based on PTAGIS metadata?
Default is `FALSE`.}
}
\description{
Build a parent-child table based on site locations and the flowlines from NHDPlusV2 layer.
}
\examples{
createParentChildDf()
}
\author{
Kevin See
}
