% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPaths.R
\name{buildPaths}
\alias{buildPaths}
\title{Build Possible Paths}
\usage{
buildPaths(parent_child = NULL)
}
\arguments{
\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function.}
}
\value{
data frame containing a column,`end_loc`, showing the possible final locations,
and a column, `path`, showing all the detection locations to pass on the way
to that final location.
}
\description{
Creates a data frame showing the path (through various detection sites) to
reach each detection location in a parent-child table.
}
\examples{
buildPaths()
}
\author{
Kevin See
}
