% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chnk2015_obs.R
\docType{data}
\name{chnk2015_obs}
\alias{chnk2015_obs}
\title{Complete tag history for spawn year 2015 valid Chinook tags

An example complete tag history observation file downloaded from PTAGIS after loading
the valid tag ID returned from running the \code{filterLGRtrapDB} and \code{summariseValidTagsLGR} functions.}
\format{A data frame with 120104 rows and 6 variables}
\source{
\url{http://www.PTAGIS.org/}
}
\usage{
chnk2015_obs
}
\keyword{datasets}
