% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCapHist.R
\name{cleanCapHist}
\alias{cleanCapHist}
\title{Remove Duplicate Detections}
\usage{
cleanCapHist(ch, config, drop_nodes = NULL)
}
\arguments{
\item{ch}{capture history file}

\item{config}{configuration file}

\item{drop_nodes}{which nodes should be dropped from capture histories?}
}
\description{
Simplify cature histories and remove multiple sequential detections at same node
}
\author{
Kevin See
}
