% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressDART.R
\name{compressDART}
\alias{compressDART}
\title{Compress Detections from DART Query}
\usage{
compressDART(
  species = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  loc = c("GRA", "PRA"),
  spawn_year = NULL,
  configuration = NULL
)
}
\arguments{
\item{species}{Chinook, Coho, Steelhead and Sockeye}

\item{loc}{code for initial location. Currently includes options for Lower Granite (\code{GRA}) and Priest Rapids (\code{PRA}) dams.}

\item{spawn_year}{available years includes spawn year 2010 to present}

\item{configuration}{is a data frame which assigns node names to unique SiteID, AntennaID, and
site configuration ID combinations. One example can be built with the function `buildConfig`. If
no configuration file is provided, nodes are considered site codes by default. If nodes are assigned,
the column name should be `node`.}
}
\value{
A tibble with particular information from a complete tag history PTAGIS query.
}
\description{
For certain locations (currently Lower Granite Dam and Priest Rapids Dam),
query DART to return all PIT tags observed there from a given species in a given spawn year,
as well as any subsequent detections upstream, and then compress them.
}
\examples{
compressDART(species = 'Chinook',
                       loc = "GRA",
                       spawn_year = 2019)
}
\author{
Kevin See
}
