% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCapHistCols.R
\name{defineCapHistCols}
\alias{defineCapHistCols}
\title{Define Column Names for Capture History Matrix}
\usage{
defineCapHistCols(
  parent_child = NULL,
  configuration = NULL,
  use_rkm = TRUE,
  bottom_sites = NULL,
  incl_all_nodes = TRUE
)
}
\arguments{
\item{parent_child}{parent-child table. Could be created from `buildParentChild()` from `PITcleanr` package.}

\item{configuration}{configuration file. Could be created from `buildConfig()` from `PITcleanr` package. If river kilometer (`rkm`) is included as a column, this function will use that to help determine the column order. Otherwise, it will use the paths contructed from `buildNodeOrder()`.}

\item{use_rkm}{if the river kilometer code exists in the configuration file (in a column named `rkm`), use those codes to help establish the order of capture history columns. Default value is `TRUE`.}

\item{bottom_sites}{If provided, represents a character vector of sites to use as "bottom sites". All upstream sites of each bottom site will be grouped just after that bottom site in the capture history matrix.}

\item{incl_all_nodes}{if some `bottom_sites` are supplied, should all the nodes in the parent-child table be included (default is `TRUE`), or only nodes upstream of those bottom sites be returned (`FALSE`). The initial site will always be returned. If no `bottom_sites` are supplied, this is set to `TRUE`.}
}
\description{
Based on a parent-child table, this returns a vector of column names to help create a consistent capture history matrix
}
\examples{
defineCapHistCols()
}
\author{
Kevin See
}
