% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineRepGrps.R
\name{defineRepGrps}
\alias{defineRepGrps}
\title{Define Report Groups}
\usage{
defineRepGrps(
  spp = c("Chinook", "Steelhead"),
  node_order = NULL,
  root_site = "LGD"
)
}
\arguments{
\item{spp}{Species, either "Chinook" or "Steelhead"}

\item{node_order}{output of function \code{createNodeOrder}.}

\item{root_site}{initial tagging site. Currently only defined for \code{LGD} (Lower Granite dam)}
}
\description{
Define which detection sites fall into which report groups, and what the lowest array is. These reporting gropus differ by species.
}
\examples{
defineRepGrps()
}
\author{
Kevin See
}
