% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFinalLoc.R
\name{estimateFinalLoc}
\alias{estimateFinalLoc}
\title{Final location}
\usage{
estimateFinalLoc(
  filtered_obs = NULL,
  incl_site_code = F,
  cth_file = NULL,
  file_type = c("PTAGIS", "Biologic_csv", "raw")
)
}
\arguments{
\item{filtered_obs}{Tibble returned by `filterDetections()`.}

\item{incl_site_code}{Should the PTAGIS site code of final location be returned? Default is `FALSE`.}

\item{cth_file}{is the path to the complete tag history file}

\item{file_type}{describes the source of the complete tag history file.

If \code{PTAGIS}, then \code{cth_file} should be a PTAGIS observation file downloaded as a csv from PTAGIS.
This must be the output from a Complete Tag History query (part of the Advanced Reporting).
This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type,
Event Site Code, Event Date Time, Antenna, Antenna Group Configuration,
Event Release Site Code, and Event Release Date Time.

If \code{Biologic_csv}, that implies the data was downloaded from Biologic software in a .csv format.

If \code{raw}, that implies the data was downloaded directly from the reader, in either a .log or .xlsx format. In this case, the largest string containing alphabetic characters in the file name will be assigned as the site code.}
}
\description{
Based on cleaned capture history, determines furthest location of each tag along possible paths.
}
\examples{
#estimateFinalLoc()
}
\author{
Kevin See
}
