% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSpawnLoc.R
\name{estimateSpawnLoc}
\alias{estimateSpawnLoc}
\title{Final spawning location}
\usage{
estimateSpawnLoc(capHist_proc = NULL)
}
\arguments{
\item{capHist_proc}{Dataframe returned by \code{processCapHist_LGD}, under the name \code{ProcCapHist}.}
}
\description{
Based on cleaned capture history, determines further upstream location of each tag. Assumption is that this is the spawning location.
}
\examples{
estimateSpawnLoc()
}
\author{
Kevin See
}
