% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSpawnLoc_LGR.R
\name{estimateSpawnLoc_LGR}
\alias{estimateSpawnLoc_LGR}
\title{Final spawning location - LGR}
\usage{
estimateSpawnLoc_LGR(filtered_obs = NULL, spawn_site = F, ptagis_file = NULL)
}
\arguments{
\item{filtered_obs}{Tibble returned by `filterDetections()`.}

\item{spawn_site}{Should the PTAGIS site code of spawning detection be returned? Default is `FALSE`.}
}
\description{
Based on cleaned capture history, determines further upstream location of each tag. Assumption is that this is the spawning location.
}
\examples{
estimateSpawnLoc_LGR()
}
\author{
Kevin See
}
