% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estNumTags.R
\name{estNumTags}
\alias{estNumTags}
\title{Estimate Tags}
\usage{
estNumTags(capHist_proc = NULL, node_order = NULL, node = NULL)
}
\arguments{
\item{capHist_proc}{Dataframe returned by \code{processCapHist_*}, under the name \code{ProcCapHist}.}

\item{node_order}{Dataframe returned by \code{processCapHist_*}, under the name \code{NodeOrder}.}

\item{node}{character string of the node(s) of interest.}
}
\description{
This function uses detections at a given node, and all detections upstream of that node to calculate detection efficiency at that node. It then uses the number of tags detected at that node and the detection efficiency to estimate the number of tags that have migrated upstream of that node.
}
\examples{
# estNumTags()
}
\author{
Kevin See
}
