% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSites.R
\name{extractSites}
\alias{extractSites}
\title{Extract Sites}
\usage{
extractSites(
  ptagis_file = NULL,
  as_sf = F,
  crs = 5070,
  min_date = NULL,
  max_date = NULL,
  configuration = NULL
)
}
\arguments{
\item{ptagis_file}{is the path to the PTAGIS observation file downloaded as a csv from PTAGIS.
This must be the output from a Complete Tag History query (part of the Advanced Reporting).
This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type,
Event Site Code, Event Date Time, Antenna, Antenna Group Configuration,
Event Release Site Code, and Event Release Date Time.}

\item{as_sf}{Should the results be returned as an `sf` object (`TRUE`), or a tibble
(`FALSE`, and the default)?}

\item{crs}{if `as_sf = TRUE`, what CRS should the resulting `sf` object be
transformed to? Default is 5070.}

\item{min_date}{character in the format `YYYYMMDD` that describes a minimum detection date.
If supplied, sites with detections prior to this date will be excluded from the results.}

\item{max_date}{character in the format `YYYYMMDD` that describes a maximum detection date.
If supplied, sites with detections after this date will be excluded from the results.}

\item{configuration}{is a data frame which assigns node names to unique SiteID, AntennaID, and
site configuration ID combinations. One example can be built with the function `buildConfig`. If
no configuration file is provided, nodes are considered site codes by default. If nodes are assigned,
the column name should be `node`.}
}
\value{
a tibble
}
\description{
The function takes complete capture histories downloaded from PTAGIS
and extracts all the detection sites from them. It matches those sites with
information from PTAGIS metadata, being sure to pull the metadata from the
correct configuration file.
}
\examples{
compress()
}
\author{
Kevin See
}
