% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTagObs.R
\name{extractTagObs}
\alias{extractTagObs}
\title{Tag Detections}
\usage{
extractTagObs(
  obs_file = NULL,
  obs_type = c("compressed", "cth"),
  file_type = c("PTAGIS", "Biologic_csv", "raw")
)
}
\arguments{
\item{obs_file}{Either a data.frame of compressed detections (output of `compress()`), or the complete tag history file path and name (same format to `cth_file` argument in `readCTH()` function).}

\item{obs_type}{what type of observation file is being fed in? `compressed` is the output of `compress()`, while `cth` is a raw complete tag history file path and name. Default is `compressed`.}

\item{file_type}{describes the source of the complete tag history file.

If \code{PTAGIS}, then \code{cth_file} should be a PTAGIS observation file downloaded as a csv from PTAGIS.
This must be the output from a Complete Tag History query (part of the Advanced Reporting).
This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type,
Event Site Code, Event Date Time, Antenna, Antenna Group Configuration,
Event Release Site Code, and Event Release Date Time.

If \code{Biologic_csv}, that implies the data was downloaded from Biologic software in a .csv format.

If \code{raw}, that implies the data was downloaded directly from the reader, in either a .log or .xlsx format. In this case, the largest string containing alphabetic characters in the file name will be assigned as the site code.}
}
\description{
List all detection sites for each tag code
}
\author{
Kevin See
}
