% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDetections.R
\name{filterDetections}
\alias{filterDetections}
\title{Determine Detections to Keep}
\usage{
filterDetections(compress_obs = NULL, parent_child = NULL, max_obs_date = NULL)
}
\arguments{
\item{compress_obs}{The result of `compress()`.}

\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function.}

\item{max_obs_date}{Character string in the format "YYYYMMDD". If included, the output will suggest that observations after this date should be deleted.}
}
\value{
a tibble
}
\description{
After determining the direction of movement using `addDirection()`,
this function provides some indication of whether each detection
should be kept for an analysis analyzing movement in a single direction
}
\examples{
filterDetections()
}
\author{
Kevin See
}
