% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLGRtrapDB.R
\name{filterLGRtrapDB}
\alias{filterLGRtrapDB}
\title{Read Trap Database}
\usage{
filterLGRtrapDB(path = ".", species = c("Chinook", "Steelhead"),
  spawnYear = NULL)
}
\arguments{
\item{path}{file path including name of file where .csv file of trap database is stored}

\item{species}{either Chinook or Steelhead}

\item{spawnYear}{spawn year}
}
\description{
Read and filter LGR adult trap database to only include fish considered part of the valid sample
}
\examples{
filterLGRtrapDB(spawnYear = 2014)
}
\author{
Kevin See
}
