% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDwnstrmHydroseg.R
\name{findDwnstrmHydroseg}
\alias{findDwnstrmHydroseg}
\title{Find Hydro Segment of Downstream Site}
\usage{
findDwnstrmHydroseg(hydro_id, flow_lines, hydro_pause_ids)
}
\arguments{
\item{hydro_id}{a character or numeric input containing the starting hydro sequence}

\item{flow_lines}{An `sf` object containing the flowlines (output of `queryFlowlines()`). Must have a column name called Hydroseq with unique identifiers, and a column name called DnHydroseq containing the hydro sequence of the next downstream segment}

\item{hydro_pause_ids}{a vector of hydro sequences containing possible end points}
}
\value{
character of the nearest downstream HY segment to starting segment that falls within a list of possible stopping segments.
}
\description{
Based on NHDPlus layer (`flow_lines`) that includes the hydro sequence of the next downstream segment, this function systematically moves downstream, identifying the next downstream hydro sequence, until it reaches a pre-determined hydro sequence number, at which point it stops.
}
\author{
Kevin See
}
