% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDwnstrmSite.R
\name{findDwnstrmSite}
\alias{findDwnstrmSite}
\title{Determine next downstream site}
\usage{
findDwnstrmSite(site_id = NULL, flow_lines = NULL, sites_joined = NULL)
}
\arguments{
\item{flow_lines}{Site code for the starting detection site.}

\item{sites_joined}{An `sf` class object containing points of all detection sites. Must contain a column named SiteID containing the site code of each site, and a column called Hydroseq containing the hydro sequence closest to each site.}
}
\value{
character of the site code downstream
}
\description{
Determine which detection site is the closest downstream site from a starting site, based on hydrographic flowlines
}
\author{
Kevin See
}
