% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeInfo.R
\name{getNodeInfo}
\alias{getNodeInfo}
\title{Compile information about each node}
\usage{
getNodeInfo(parent_child = NULL, configuration = NULL)
}
\arguments{
\item{parent_child}{parent-child table. Could be created from `buildParentChild()` from `PITcleanr` package.}

\item{configuration}{configuration file. Could be created from `buildConfig()` from `PITcleanr` package.}
}
\description{
Based on a parent-child table and a configuration file, this function compiles how many nodes each site has, what the parent site is, and what child number of that parent the site in question is
}
\examples{
#getNodeInfo()
}
\author{
Kevin See
}
