% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidPaths.R
\name{getValidPaths}
\alias{getValidPaths}
\title{Valid Paths}
\usage{
getValidPaths(parent_child_df = NULL, root_site = NULL)
}
\arguments{
\item{parent_child_df}{dataframe with at least \code{ParentNode} and \code{ChildNode} columns}
}
\description{
Construct data.frame of valid paths to each node
}
\examples{
getValidPaths()
}
\author{
Kevin See
}
