% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listChildNodes.R
\name{listChildNodes}
\alias{listChildNodes}
\title{Child Nodes}
\usage{
listChildNodes(parent_node, parent_child_df)
}
\arguments{
\item{parent_node}{Site code from parent node}
}
\description{
Recursive function for finding all child nodes of parent node
}
\examples{
listChildNodes()
}
\author{
Kevin See
}
