% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listChildSites.R
\name{listChildSites}
\alias{listChildSites}
\title{Child Sites}
\usage{
listChildSites(parent_site, parent_child_df)
}
\arguments{
\item{parent_site}{Site code from parent}
}
\description{
Recursive function for finding all child sites of parent site
}
\examples{
listChildSites()
}
\author{
Kevin See
}
