% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listParentNodes.R
\name{listParentNodes}
\alias{listParentNodes}
\title{Parent Nodes}
\usage{
listParentNodes(end_node, parent_child_df, root_site = "GRA")
}
\arguments{
\item{end_node}{Code of final node}

\item{parent_child_df}{dataframe with at least \code{ParentNode} and \code{ChildNode} columns}

\item{root_site}{initial site; dictates how far backwards this function should search. Default value is \code{GRA} for Lower Granite Dam.}
}
\description{
Recursive function for finding all parent nodes of child node
}
\examples{
listParentNodes()
}
\author{
Kevin See
}
