% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listParents.R
\name{listParents}
\alias{listParents}
\title{List Parent Locations}
\usage{
listParents(loc_code = NULL, parent_child = NULL)
}
\arguments{
\item{loc_code}{Code of child location, contained in `parent_child`}

\item{parent_child}{dataframe with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function.}
}
\value{
character vector of all sites leading to `loc_code` site.
}
\description{
Recursive function for finding all parents of child location
}
\examples{
listParents()
}
\author{
Kevin See
}
