% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeDetectionEvent.R
\name{nodeDetectionEvent}
\alias{nodeDetectionEvent}
\title{Node Detection Event Sequencing}
\usage{
nodeDetectionEvent(x)
}
\arguments{
\item{x}{character string of Node names}
}
\value{
character string
}
\description{
the function uses base R and run length encoding to identify the
  number of times a fish is detected on a single node during a single event and before being interupted
  with a detection on a different node.  The function returns a character string the same length as \code{x}.
}
\examples{
nodeDetectionEvent()
}
\author{
Ryan N. Kinzer
}
