% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsNetwork.R
\name{obsNetwork}
\alias{obsNetwork}
\title{Observation Network}
\usage{
obsNetwork(parentchild)
}
\arguments{
\item{parentchild}{a dataframe containing at least two columns; 1: parent and 2: child for each observation point in the network.}
}
\description{
Create data.frame describing the observation network and paths from a parent/child dataset.
}
\examples{
obsNetwork()
}
\author{
Ryan N. Kinzer
}
