% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNodes.R
\name{plotNodes}
\alias{plotNodes}
\title{Detection Node Schematic}
\usage{
plotNodes(
  parent_child = NULL,
  layout = "auto",
  point_size = 5,
  label_points = T,
  label_size = 2,
  ...
)
}
\arguments{
\item{parent_child}{The output of `buildParentChild()`.}

\item{layout}{The type of layout to create. Either a valid string, a
function, a matrix, or a data.frame (see Details)}

\item{point_size}{size for points on graph (passed to `size` argument of `geom_node_point()`).}

\item{label_points}{`TRUE/FALSE` should the nodes be labeled with their names? Default is `TRUE`.}

\item{label_size}{size for node labels on graph (passed to `size` argument of `geom_node_label()`).}

\item{...}{other arguments for `geom_edge_diagonal`, `geom_node_point()` or `geom_node_label()`.}
}
\description{
Create a schematic plot showing the connections
between each detection node in a parent-child table.
}
\author{
Kevin See
}
