% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCapHist_PRD.R
\name{processCapHist_PRD}
\alias{processCapHist_PRD}
\title{Process capture histories for Priest Rapids}
\usage{
processCapHist_PRD(
  startDate = NULL,
  configuration = NULL,
  parent_child = NULL,
  observations = NULL,
  last_obs_date = NULL,
  truncate = T,
  site_df,
  step_num = 1,
  save_file = F,
  file_name = NULL
)
}
\arguments{
\item{startDate}{observations that occurred before this date (YYYYMMDD format) will not be included.}

\item{last_obs_date}{Character string in the format "YYYYMMDD". If included, the output will suggest that observations after this date should be deleted.}

\item{save_file}{Should output be written to an file? Default value is \code{FALSE}.}

\item{file_name}{If \code{save_file == TRUE}, this is the file name (with possible extension) to be saved to.}
}
\description{
Using a site/node configuration file, a list of modeled sites, a parent-child table of nodes, and the raw capture histories from PTAGIS, process them. This function is specific to fish tagged at Priest Rapids dam.
}
\examples{
processCapHist_PRD()
}
\author{
Kevin See
}
