% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcTagHistory.R
\name{qcTagHistory}
\alias{qcTagHistory}
\title{Quality Check Complete Tag Histories}
\usage{
qcTagHistory(
  cth_file = NULL,
  file_type = c("PTAGIS", "Biologic_csv", "raw"),
  ignore_event_vs_release = F
)
}
\arguments{
\item{cth_file}{is the path to the complete tag history file}

\item{file_type}{describes the source of the complete tag history file.

If \code{PTAGIS}, then \code{cth_file} should be a PTAGIS observation file downloaded as a csv from PTAGIS.
This must be the output from a Complete Tag History query (part of the Advanced Reporting).
This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type,
Event Site Code, Event Date Time, Antenna, Antenna Group Configuration,
Event Release Site Code, and Event Release Date Time.

If \code{Biologic_csv}, that implies the data was downloaded from Biologic software in a .csv format.

If \code{raw}, that implies the data was downloaded directly from the reader, in either a .log or .xlsx format. In this case, the largest string containing alphabetic characters in the file name will be assigned as the site code.}

\item{ignore_event_vs_release}{Should the function attempt to choose whether to use the event time value
or the event release time value for different release batches? Default is `FALSE`. If set to `TRUE`,
the event time value will be used}
}
\value{
a list consisting of a vector of tag codes that are listed as "Disown" in PTAGIS (`disown_tags`),
a vector of tag codes that are listed as "Orphan" in PTAGIS (`orphan_tags`), and,
if `ignore_event_vs_release` is `TRUE`,
a tibble containing information about sites with release information to help the user
determine with event time or release time should be used for those sites.
}
\description{
The function takes complete capture histories downloaded from PTAGIS
and performs several QA/QC checks on them. It returns the results in the form of a list.
}
\examples{
cth_file = system.file("extdata", "PRO_Steelhead_2019.csv", package = "PITcleanr")
qcTagHistory(cth_file)
}
\author{
Kevin See
}
