% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFlowlines.R
\name{queryFlowlines}
\alias{queryFlowlines}
\title{Download NHDPlus flow lines}
\usage{
queryFlowlines(
  sites_sf = NULL,
  root_site_code = NULL,
  min_strm_order = 0,
  max_upstream_comid = NULL,
  set_crs = NULL
)
}
\arguments{
\item{sites_sf}{An `sf` class object containing points of all detection sites. Must contain a column named `site_code` containing the site code of each site.}

\item{root_site_code}{Site code for the starting detection site.}

\item{min_strm_order}{minimum stream order to query flowlines for. Default value is `0`.}

\item{max_upstream_comid}{if provided, this function will filter out any flowlines upstream of this unique identifier (COMID) in the NHDPlus stream network. Default value is `NULL`. For assistance identifying a COMID, please see the `hydrogeofetch` package, specifically the `discover_nhdplus_id()` function.}

\item{set_crs}{What CRS should be used? Default will read the CRS from the `sites_sf` object.}
}
\value{
list with at least two elements: `flowlines` is the `sf` object of flowlines downloaded from NHDPlus. `basin` is the polygon representing the catchment upstream of the root site.
}
\description{
Download a subset of the NHDPlus v2.1 flowlines that covers a list of detection sites.
}
\author{
Kevin See
}
