% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryInterrogationSites.R
\name{queryInterrogationSites}
\alias{queryInterrogationSites}
\title{PTAGIS Interagation sites}
\source{
\url{http://www.ptagis.org}
}
\usage{
queryInterrogationSites(site = NULL)
}
\arguments{
\item{site}{PTAGIS site code. Default is \code{NULL} which will query all sites}
}
\description{
Query and download metadata for PTAGIS Interrogation sites
}
\examples{
queryInterrogationSites()
}
\author{
Kevin See
}
