% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryMRRDataFile.R
\name{queryMRRDataFile}
\alias{queryMRRDataFile}
\title{PTAGIS MRR Data File}
\source{
\url{http://www.ptagis.org}
}
\usage{
queryMRRDataFile(file_nm = NULL, file_path = NULL)
}
\arguments{
\item{file_nm}{PTAGIS file name.}

\item{file_path}{if the user has downloaded the PTAGIS file, this function can read in the file, if the file path is included here. The default value, \code{NULL}, indicates that the file should be queried from the PTAGIS API interface.}
}
\description{
Queries a raw MRR data file for PTAGIS MRR sites, based upon a known file name. NOTE: only the latest, corrected, version of a MRR file is returned
}
\examples{
queryMRRDataFile(file_nm = "NBD-2014-188-PRD.xml")
}
\author{
Kevin See
}
