% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryMRRSites.R
\name{queryMRRSites}
\alias{queryMRRSites}
\title{PTAGIS MRR sites}
\source{
\url{http://www.ptagis.org}
}
\usage{
queryMRRSites(site = NULL)
}
\arguments{
\item{site}{PTAGIS site code. Default is \code{NULL} which will query all sites}
}
\description{
Query and download metadata for PTAGIS MRR sites
}
\examples{
queryMRRSites()
}
\author{
Kevin See
}
