% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryObsDART.R
\name{queryObsDART}
\alias{queryObsDART}
\title{Query DART for observations of PIT tags}
\usage{
queryObsDART(
  species = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  loc = c("GRA", "PRA"),
  spawn_year = NULL
)
}
\arguments{
\item{species}{Chinook, Coho, Steelhead and Sockeye}

\item{loc}{code for initial location. Currently includes options for Lower Granite (\code{GRA}) and Priest Rapids (\code{PRA}) dams.}

\item{spawn_year}{available years includes spawn year 2010 to present}
}
\value{
A tibble with particular information from a complete tag history PTAGIS query.
}
\description{
For certain locations (currently Lower Granite Dam and Priest Rapids Dam),
query DART to return all PIT tags observed there from a given species in a given spawn year,
as well as any subsequent detections upstream.
}
\examples{
queryObsDART(species = 'Chinook',
                       loc = "GRA",
                       spawn_year = 2019)
}
\author{
Kevin See
}
