% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPtagisMeta.R
\name{queryPtagisMeta}
\alias{queryPtagisMeta}
\title{Query PTAGIS Site Metadata}
\source{
\url{http://www.ptagis.org}
}
\usage{
queryPtagisMeta()
}
\description{
Compile metadata from all MRR and interogation sites from PTAGIS
}
\examples{
queryPtagisMeta()

}
\author{
Kevin See
}
