% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryTagMeta.R
\name{queryTagMeta}
\alias{queryTagMeta}
\title{PTAGIS MRR sites}
\source{
\url{http://www.ptagis.org}
}
\usage{
queryTagMeta(tagCode = NULL)
}
\arguments{
\item{tagCode}{PTAGIS tag code}
}
\description{
Query and download mark data for one tag from PTAGIS
}
\examples{
queryTagMeta()
}
\author{
Kevin See
}
