% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryTagsDART.R
\name{queryTagsDART}
\alias{queryTagsDART}
\title{Query DART for list of PIT tags}
\usage{
queryTagsDART(
  species = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  loc = c("GRA", "PRA"),
  spawnYear = NULL
)
}
\arguments{
\item{species}{Chinook, Coho, Steelhead and Sockeye}

\item{loc}{code for initial location. Currently includes options for Lower Granite (\code{GRA}) and Priest Rapids (\code{PRA}) dams.}

\item{spawnYear}{available years includes spawn year 2010 to present}
}
\description{
For certain locations, query DART to return all PIT tags observed there from a given species in a given spawn year.
}
\examples{
queryTagsDART("Chinook", "GRA", 2019)

}
\author{
Kevin See
}
