% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCTH.R
\name{readCTH}
\alias{readCTH}
\title{Read In PTAGIS Complete Tag History}
\usage{
readCTH(ptagis_file = NULL)
}
\arguments{
\item{ptagis_file}{is the path to the PTAGIS observation file downloaded as a csv from PTAGIS.
This must be the output from a Complete Tag History query (part of the Advanced Reporting).
This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type,
Event Site Code, Event Date Time, Antenna, Antenna Group Configuration,
Event Release Site Code, and Event Release Date Time.}
}
\value{
a list
}
\description{
The function reads in complete tag histories from PTAGIS that
have been downloaded and coverts the date time values from character
vectors into date-time vectore
}
\examples{
qcTagHistory()
}
\author{
Kevin See
}
