% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTestTag.R
\name{readTestTag}
\alias{readTestTag}
\title{Read In Test Tag Data}
\usage{
readTestTag(
  cth_file = NULL,
  file_type = c("PTAGIS", "Biologic_csv", "raw"),
  test_tag_prefix = "3E7"
)
}
\arguments{
\item{cth_file}{is the path to the complete tag history file}

\item{file_type}{describes the source of the complete tag history file.

If \code{PTAGIS}, then \code{cth_file} should be a PTAGIS observation file downloaded as a csv from PTAGIS.
This must be the output from a Complete Tag History query (part of the Advanced Reporting).
This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type,
Event Site Code, Event Date Time, Antenna, Antenna Group Configuration,
Event Release Site Code, and Event Release Date Time.

If \code{Biologic_csv}, that implies the data was downloaded from Biologic software in a .csv format.

If \code{raw}, that implies the data was downloaded directly from the reader, in either a .log or .xlsx format. In this case, the largest string containing alphabetic characters in the file name will be assigned as the site code.}

\item{test_tag_prefix}{The prefix that designates a tag code as a test tag. These detections are filtered out of the returned tibble}
}
\value{
a tibble containing the raw detections of test tags
}
\description{
The function reads in complete tag histories and filters out any tag not designated a
test tag by the tag code prefix.
}
\examples{
ptagis_file = system.file("extdata", "PRO_Steelhead_2019.csv", package = "PITcleanr")
readTestTag(ptagis_file)
}
\author{
Kevin See
}
