% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTagData.R
\name{summariseTagData}
\alias{summariseTagData}
\title{Tag Summary}
\usage{
summariseTagData(
  capHist_proc = NULL,
  trap_data = NULL,
  saveCSV = F,
  file_name = NULL
)
}
\arguments{
\item{capHist_proc}{Dataframe returned by \code{processCapHist_LGD}, under the name \code{ProcCapHist}.}

\item{trap_data}{Dataframe containing tag ID and relevant biological data from the adult trap. Part of the list returned by \code{processCapHist_LGD}, under the name \code{ValidTrapData}.}

\item{saveCSV}{should the resulting dataframe be saved as a csv file? Default value is \code{FALSE}.}

\item{file_name}{if \code{saveCSV} is \code{TRUE}, this is the file path and name, with .csv extension, where the file should be save.}
}
\description{
For each tag, summarises final spawning location as well as important pieces of biological information.
}
\examples{
summariseTagData()
}
\author{
Kevin See
}
