% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseValidTagsLGR.R
\name{summariseValidTagsLGR}
\alias{summariseValidTagsLGR}
\title{Summarise Valid Tags}
\usage{
summariseValidTagsLGR(valid_df)
}
\arguments{
\item{valid_df}{dataframe of valid tags extracted from the trap database with the \code{filterLGRtrapDB} function}
}
\description{
Filter the valid tag dataframe returned by \code{filterLGRtrapDB} to only include the tag code and the tagging date
}
\examples{
summariseValidTagsLGR(filterLGRtrapDB(spawnYear = 2015))
}
\author{
Kevin See
}
