% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTagData.R
\name{summarizeTagData}
\alias{summarizeTagData}
\title{Tag Summary}
\usage{
summarizeTagData(
  filtered_obs = NULL,
  bio_data = NULL,
  saveCSV = F,
  file_name = NULL,
  ...
)
}
\arguments{
\item{filtered_obs}{Tibble returned by `filterDetections()`.}

\item{bio_data}{Dataframe containing `tag_code` and relevant biological data from each fish, often gathered when they were tagged.}

\item{file_name}{if `save_csv` is `TRUE`, this is the file path and name, with .csv extension, where the file should be save.}

\item{save_csv}{should the resulting dataframe be saved as a csv file? Default value is `FALSE`.}
}
\description{
For each tag, summarises final spawning location as well as important pieces of biological information.
}
\examples{
summarizeTagData()
}
\author{
Kevin See
}
