% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_chnk2015.R
\docType{data}
\name{trap_chnk2015}
\alias{trap_chnk2015}
\title{Sample of adult trap database

A sample of the Lower Granite adult fish trap database containing data on spring/summer Chinook that crossed the dam in 2015.}
\format{A data frame with 12847 rows and 52 variables}
\usage{
trap_chnk2015
}
\keyword{datasets}
