% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCapHistOutput.R
\name{writeCapHistOutput}
\alias{writeCapHistOutput}
\title{Capture History Output}
\usage{
writeCapHistOutput(valid_obs = NULL, valid_paths = NULL,
  node_order = NULL, save_file = F, file_name = NULL)
}
\arguments{
\item{valid_obs}{dataframe built by the function \code{assignNodes}.}

\item{valid_paths}{dataframe built by the function \code{getValidPaths}.}

\item{node_order}{dataframe built by the function \code{createNodeOrder}.}

\item{save_file}{Should output be written to an Excel file? Default value is \code{FALSE}.}

\item{file_name}{If \code{save_file == TRUE}, this is the file name (with possible extension) to be saved to.}
}
\description{
This function combines the results from \code{writeFishPaths} and \code{writeSpwnPaths} functions and saves it as an excel file if desired.
}
\author{
Kevin See
}
