% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFishPaths.R
\name{writeFishPaths}
\alias{writeFishPaths}
\title{Fish Paths}
\usage{
writeFishPaths(valid_obs, valid_paths)
}
\arguments{
\item{valid_obs}{}

\item{valid_paths}{}
}
\description{
The function builds a data frame of all the observed fish paths in the
observation file based on the parent child table. The function was originally called
FishPath_truncateToLastLeg() and was built by Greg K. Ryan K. altered the function
to work solely in R and doesn't require the SQLite backend, and called in fishPaths(). Kevin S. edited it further
}
\examples{
writeFishPaths()
}
\author{
Greg Kliewer, Ryan Kinzer, Kevin See
}
