% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeOldNetworks.R
\name{writeOldNetworks}
\alias{writeOldNetworks}
\title{Old Node Networks}
\usage{
writeOldNetworks()
}
\description{
Calls old functions to create data.frames describing all the observation nodes in various versions of DABOM, including how they are related to one another. Kept for backwards compatibility checks.
}
\examples{
writeLGRNodeNetwork()
}
\author{
Kevin See
}
