% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePRDNodeNetwork.R
\name{writePRDNodeNetwork}
\alias{writePRDNodeNetwork}
\title{Priest Dam Node Network}
\usage{
writePRDNodeNetwork()
}
\description{
Create data.frame describing all the observation nodes in the Priest dam version of DABOM, including how they are related to one another
}
\examples{
writePRDNodeNetwork()
}
\author{
Kevin See
}
