% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTUMNodeNetwork.R
\name{writeTUMNodeNetwork}
\alias{writeTUMNodeNetwork}
\title{Tumwater Node Network}
\usage{
writeTUMNodeNetwork()
}
\description{
Create data.frame describing all the observation nodes in the Tumwater dam version of DABOM, including how they are related to one another
}
\examples{
writeTUMNodeNetwork()
}
\author{
Kevin See
}
