% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTUMNodeNetwork_noUWE.R
\name{writeTUMNodeNetwork_noUWE}
\alias{writeTUMNodeNetwork_noUWE}
\title{Tumwater Node Network}
\usage{
writeTUMNodeNetwork_noUWE()
}
\description{
Create data.frame describing all the observation nodes in the Tumwater dam version of DABOM, including how they are related to one another. This function does not include the detection site UWE
}
\examples{
writeTUMNodeNetwork_noUWE()
}
\author{
Kevin See
}
