% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildConfig.R
\name{buildConfig}
\alias{buildConfig}
\title{Build configuration file}
\source{
\url{http://www.ptagis.org}
}
\usage{
buildConfig(node_assign = c("array", "site", "antenna"))
}
\arguments{
\item{node_assign}{should nodes by assigned by \code{array}, \code{site} or \code{antenna}? Default is \code{array}.}
}
\description{
Compile metadata from all MRR and interogation sites from PTAGIS
}
\examples{
buildConfig()

}
\author{
Kevin See
}
