% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildConfig.R
\name{buildConfig}
\alias{buildConfig}
\title{Build configuration file}
\source{
\url{http://www.ptagis.org}
}
\usage{
buildConfig()
}
\description{
Compile metadata from all MRR and interogation sites from PTAGIS
}
\examples{
buildConfig()

}
\author{
Kevin See
}
