% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNodeOrder.R
\name{buildNodeOrder}
\alias{buildNodeOrder}
\title{Build Node Order}
\usage{
buildNodeOrder(parent_child = NULL, direction = "u")
}
\arguments{
\item{parent_child}{data frame with at least `parent` and `child` columns.
Can be created with `buildParentChild()` function.}
}
\value{
data frame containing a column,`end_loc`, showing the possible final locations,
a column, `node_order`, showing the order of that node along the path,
and a column, `path`, showing all the detection locations to pass on the way
to that final location.
}
\description{
Creates a data frame showing the path (through various detection sites) to
reach each detection location in a parent-child table, and the node order of that location
}
\examples{
buildNodeOrder()
}
\author{
Kevin See
}
