% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNodeOrder.R
\name{createNodeOrder}
\alias{createNodeOrder}
\title{Node Order}
\usage{
createNodeOrder(valid_paths = NULL, configuration = NULL,
  site_df = NULL, step_num = 1)
}
\arguments{
\item{valid_paths}{a dataframe created by the \code{getValidPaths} function.}

\item{configuration}{a configuration file built by \code{buildConfig} function, and possibly modified by the user.}

\item{site_df}{Optional. A data frame created by one of the \code{write___NodeNetwork}.}

\item{step_num}{in the \code{site_df} data frame, which step should be used to group various nodes? Default value is 1.}
}
\description{
Based on valid paths, returns a dataframe with a row for each node, showing the path to get to that node, and how many nodes a tag must pass to get there.
}
\examples{
#createNodeOrder()
}
\author{
Kevin See
}
