% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editParentChild.R
\name{editParentChild}
\alias{editParentChild}
\title{Edit a Parent-Child table}
\source{
\url{http://www.ptagis.org}
}
\usage{
editParentChild(
  parent_child = NULL,
  parent_locs = NULL,
  child_locs = NULL,
  new_parent_locs = NULL
)
}
\arguments{
\item{parent_child}{dataframe produced by `buildParentChild()`.}

\item{new_parent_locs}{}
}
\description{
Allow the user to easily make manual changes to a parent child table.
}
\examples{
createParentChildDf()
}
\author{
Kevin See
}
