% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSpawnLoc.R
\name{estimateSpawnLoc}
\alias{estimateSpawnLoc}
\title{Final spawning location}
\usage{
estimateSpawnLoc(
  filtered_obs = NULL,
  spawn_site = F,
  cth_file = NULL,
  file_type = c("PTAGIS", "Biologic_csv", "raw")
)
}
\arguments{
\item{filtered_obs}{Tibble returned by `filterDetections()`.}

\item{spawn_site}{Should the PTAGIS site code of spawning detection be returned? Default is `FALSE`.}

\item{cth_file}{is the path to the complete tag history file}

\item{file_type}{describes the source of the complete tag history file.

If \code{PTAGIS}, then \code{cth_file} should be a PTAGIS observation file downloaded as a csv from PTAGIS.
This must be the output from a Complete Tag History query (part of the Advanced Reporting).
This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type,
Event Site Code, Event Date Time, Antenna, Antenna Group Configuration,
Event Release Site Code, and Event Release Date Time.

If \code{Biologic_csv}, that implies the data was downloaded from Biologic software in a .csv format.

If \code{raw}, that implies the data was downloaded directly from the reader, in either a .log or .xlsx format. In this case, the largest string containing alphabetic characters in the file name will be assigned as the site code.}
}
\description{
Based on cleaned capture history, determines further upstream location of each tag. Assumption is that this is the spawning location.
}
\examples{
estimateSpawnLoc()
}
\author{
Kevin See
}
