% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSites.R
\name{extractSites}
\alias{extractSites}
\title{Extract Sites}
\usage{
extractSites(ptagis_file = NULL, as_sf = F, crs = 5070)
}
\arguments{
\item{ptagis_file}{is the path to the PTAGIS observation file downloaded as a csv from PTAGIS.
This must be the output from a Complete Tag History query (part of the Advanced Reporting).
This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type,
Event Site Code, Event Date Time, Antenna, Antenna Group Configuration,
Event Release Site Code, and Event Release Date Time.}

\item{as_sf}{Should the results be returned as an `sf` object (`TRUE`), or a tibble
(`FALSE`, and the default)?}

\item{crs}{if `as_sf = TRUE`, what CRS should the resulting `sf` object be
transformed to? Default is 5070.}
}
\value{
a tibble
}
\description{
The function takes complete capture histories downloaded from PTAGIS
and extracts all the detection sites from them. It matches those sites with
information from PTAGIS metadata, being sure to pull the metadata from the
correct configuration file.
}
\examples{
compress()
}
\author{
Kevin See
}
