% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTagObs.R
\name{extractTagObs}
\alias{extractTagObs}
\title{Tag Detections}
\usage{
extractTagObs(ptagis_file = NULL)
}
\arguments{
\item{ptagis_file}{is the path to the PTAGIS observation file downloaded as a csv from PTAGIS.
This must be the output from a Complete Tag History query (part of the Advanced Reporting).
This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type,
Event Site Code, Event Date Time, Antenna, Antenna Group Configuration,
Event Release Site Code, and Event Release Date Time.}
}
\description{
List all detection sites for each tag code
}
\author{
Kevin See
}
