% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLGRtrapDB.R
\name{filterLGRtrapDB}
\alias{filterLGRtrapDB}
\title{Read Lower Granite Trap Database}
\usage{
filterLGRtrapDB(trap_path = NULL, species = c("Chinook", "Steelhead"),
  spawnYear = NULL, saveValidTagList = F, validTagFileNm = NULL,
  saveCSV = F, validTagDataFileNm = NULL)
}
\arguments{
\item{trap_path}{file path including name of file where .csv file of trap database is stored}

\item{species}{either Chinook or Steelhead}

\item{spawnYear}{spawn year, as integer value, e.g. \code{2015}}

\item{saveValidTagList}{Should text file of valid tags be saved in a format able to be uploaded to PTAGIS for a batch query? Default value is \code{FALSE}.}

\item{validTagFileNm}{if \code{saveValidTagList} is \code{TRUE}, this is the path and filename (including .txt) to save that file as.}

\item{saveCSV}{Should csv file of all biological information from the trapping database for the valid tags be saved? Default value is \code{FALSE}.}

\item{validTagDataFileNm}{if \code{saveCSV} is \code{TRUE}, this is the path and filename (including .csv) to save that file as.}
}
\description{
Read and filter LGR adult trap database to only include fish considered part of the valid sample for a given species (spring/summer Chinook salmon or steelhead) and given spawn year.
}
\examples{
filterLGRtrapDB(trap_path = NULL, species == 'Chinook', spawnYear = 2015)
}
\author{
Kevin See, Mike Ackerman, Ryan Kinzer, Rick Orme
}
