% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCapHist_LGD.R
\name{processCapHist_LGD}
\alias{processCapHist_LGD}
\title{Process capture histories for Lower Granite}
\usage{
processCapHist_LGD(species = c("Chinook", "Steelhead"),
  spawnYear = NULL, configuration = NULL, parent_child = NULL,
  trap_path = NULL, filter_by_PBT = T, observations = NULL,
  last_obs_date = NULL, truncate = T, site_df, step_num = 1,
  save_file = F, file_name = NULL)
}
\arguments{
\item{species}{either Chinook or Steelhead}

\item{spawnYear}{spawn year, as integer value, e.g. \code{2015}}

\item{configuration}{is a data frame which assigns node names to unique SiteID, AntennaID, and site configuration ID combinations. It can be built with the function \code{buildConfig}.}

\item{trap_path}{file path including name of file where .csv file of trap database is stored}

\item{filter_by_PBT}{Should fish identified as hatchery origin be filtered out of the sample, based on PBT results? Default value is \code{TRUE}.}

\item{last_obs_date}{Character string in the format "YYYYMMDD". If included, the output will suggest that observations after this date should be deleted.}

\item{truncate}{logical, subsets observations to those with valid nodes, observations dates greater than trapping date at LGD and then to the minimum observation date of each set of observation events at a node, multiple observation events can occur at one node if the observations are split by detections at other nodes. Default is \code{TRUE}.}

\item{site_df}{Optional. A data frame created by one of the \code{write___NodeNetwork}.}

\item{step_num}{in the \code{site_df} data frame, which step should be used to group various nodes? Default value is 1.}

\item{save_file}{Should output be written to an file? Default value is \code{FALSE}.}

\item{file_name}{If \code{save_file == TRUE}, this is the file name (with possible extension) to be saved to.}
}
\description{
Using a site/node configuration file, a list of modeled sites, a parent-child table of nodes, and the raw capture histories from PTAGIS, process them. This function is specific to fish tagged at Lower Granite dam.
}
\examples{
processCapHist_LGD()
}
\author{
Kevin See
}
