% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCapHist_TUM.R
\name{processCapHist_TUM}
\alias{processCapHist_TUM}
\title{Process capture histories for Tumwater}
\usage{
processCapHist_TUM(species = c("Chinook", "Steelhead"),
  spawnYear = NULL, configuration = NULL, parent_child = NULL,
  observations = NULL, last_obs_date = NULL, truncate = T,
  save_file = F, file_name = NULL)
}
\arguments{
\item{configuration}{is a data frame which assigns node names to unique SiteID, AntennaID, and site configuration ID combinations. It can be built with the function \code{buildConfig}.}

\item{last_obs_date}{Character string in the format "YYYYMMDD". If included, the output will suggest that observations after this date should be deleted.}

\item{truncate}{logical, subsets observations to those with valid nodes, observations dates greater than trapping date at LGD and then to the minimum observation date of each set of observation events at a node, multiple observation events can occur at one node if the observations are split by detections at other nodes. Default is \code{TRUE}.}

\item{save_file}{Should output be written to an file? Default value is \code{FALSE}.}

\item{file_name}{If \code{save_file == TRUE}, this is the file name (with possible extension) to be saved to.}
}
\description{
Using a site/node configuration file, a list of modeled sites, a parent-child table of nodes, and the raw capture histories from PTAGIS, process them. This function is specific to fish tagged at Tumwater dam.
}
\examples{
processCapHist_TUM()
}
\author{
Kevin See
}
