% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDART_LGR.R
\name{processDART_LGR}
\alias{processDART_LGR}
\title{Process tags observed at Lower Granite and any subsequent detections upstream.}
\usage{
processDART_LGR(
  species = c("Chinook", "Steelhead"),
  spawnyear = NULL,
  observations = NULL,
  configuration = NULL,
  truncate = T
)
}
\description{
Using the raw import from DART's observation query, and the configuration files the function process the data into a PITcleanr capture history dataframe.
}
\examples{
processDART_LGR()
}
\author{
Ryan N. Kinzer
}
