% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcTagHistory.R
\name{qcTagHistory}
\alias{qcTagHistory}
\title{Quality Check Complete Tag Histories}
\usage{
qcTagHistory(ptagis_file = NULL)
}
\arguments{
\item{ptagis_file}{is the path to the PTAGIS observation file downloaded as a csv from PTAGIS.
This must be the output from a Complete Tag History query (part of the Advanced Reporting).
This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type,
Event Site Code, Event Date Time, Antenna, Antenna Group Configuration,
Event Release Site Code, and Event Release Date Time.}
}
\value{
a list
}
\description{
The function takes complete capture histories downloaded from PTAGIS
and performs several QA/QC checks on them. It returns the results in the form of a list.
}
\examples{
qcTagHistory()
}
\author{
Kevin See
}
