% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCapHist.R
\name{queryCapHist}
\alias{queryCapHist}
\title{PTAGIS Detection History}
\source{
\url{http://www.ptagis.org}
}
\usage{
queryCapHist(tag_code = NULL, configuration = NULL)
}
\arguments{
\item{tag_code}{PTAGIS tag code}
}
\description{
Query and download complete capture history for one tag from PTAGIS (through DART). Due to
PTAGIS no longer offering queries for the mark event information for a given tag code, only
observations after the mark event are returned.
}
\examples{
queryCapHist(tag_code = '384.3B23A0A6EE')
}
\author{
Kevin See
}
