% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFlowlines.R
\name{queryFlowlines}
\alias{queryFlowlines}
\title{Download NHDPlus flow lines}
\usage{
queryFlowlines(
  sites_sf = NULL,
  root_site_code = NULL,
  dwnstrm_sites = FALSE,
  dwn_min_stream_order_diff = NULL
)
}
\arguments{
\item{sites_sf}{An `sf` class object containing points of all detection sites. Must contain a column named SiteID containing the site code of each site.}

\item{root_site_code}{Site code for the starting detection site.}
}
\value{
sf
}
\description{
Download a subset of the NHDPlus v2.1 flowlines that covers a list of detection sites.
}
\author{
Kevin See
}
