% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryInterrogrationConfig.R
\name{queryInterrogationConfig}
\alias{queryInterrogationConfig}
\title{PTAGIS Interagation sites configuration}
\source{
\url{http://www.ptagis.org}
}
\usage{
queryInterrogationConfig(site = NULL)
}
\arguments{
\item{site}{PTAGIS site code. Default is \code{NULL} which will query all sites}
}
\description{
Query and download configuration metadata for PTAGIS Interrogation sites
}
\examples{
queryInterrogationConfig(site = "ZEN")
}
\author{
Kevin See
}
