% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryInterrogationMeta.R
\name{queryInterrogationMeta}
\alias{queryInterrogationMeta}
\title{PTAGIS Interagation sites metadata}
\source{
\url{http://www.ptagis.org}
}
\usage{
queryInterrogationMeta(site = NULL)
}
\arguments{
\item{site}{PTAGIS site code. Default is \code{NULL} which will query all sites}
}
\description{
Query and download metadata for PTAGIS Interrogation sites
}
\examples{
queryInterrogationMeta(site = 'ZEN')
}
\author{
Kevin See
}
