% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryObsDART.R
\name{queryObsDART}
\alias{queryObsDART}
\title{Query DART for observations of PIT tags}
\usage{
queryObsDART(
  species = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  loc = c("GRA", "PRA"),
  spawnYear = NULL
)
}
\arguments{
\item{species}{Chinook, Coho, Steelhead and Sockeye}

\item{loc}{code for initial location. Currently includes options for Lower Granite (\code{GRA}) and Priest Rapids (\code{PRA}) dams.}

\item{spawnYear}{available years includes spawn year 2010 to present}
}
\description{
For certain locations, query DART to return all PIT tags observed there from a given species in a given spawn year, as well as any subsequent detections upstream.
}
\examples{

my_config <- buildConfig()
dart_obs <- queryObsDART(species = 'Chinook',
                           spawnYear = 2020,
                           configuration = my_config,
                           truncate = T)
}
\author{
Kevin See
}
