% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryTagMeta.R
\name{queryTagMeta}
\alias{queryTagMeta}
\title{PTAGIS MRR sites}
\source{
\url{http://www.ptagis.org}
}
\usage{
queryTagMeta(tagCode = NULL)
}
\arguments{
\item{tagCode}{PTAGIS tag code}
}
\value{
From PTAGIS: WARNING: this feature has been disabled. Contact PTAGIS for more info. Gets mark event information for a given tag code.
}
\description{
Query and download mark data for one tag from PTAGIS
}
\examples{
queryTagMeta("3D9.1C2D929849")
}
\author{
Kevin See
}
