% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCTH.R
\name{readCTH}
\alias{readCTH}
\title{Read In Complete Tag History}
\usage{
readCTH(cth_file = NULL, file_type = c("PTAGIS", "Biologic_csv", "raw"))
}
\arguments{
\item{cth_file}{is the path to the complete tag history file}

\item{file_type}{describes the source of the complete tag history file.

If \code{PTAGIS}, then \code{cth_file} should be a PTAGIS observation file downloaded as a csv from PTAGIS.
This must be the output from a Complete Tag History query (part of the Advanced Reporting).
This query should contain: Tag, Mark Species, Mark Rear Type, Event Type, Event Site Type,
Event Site Code, Event Date Time, Antenna, Antenna Group Configuration,
Event Release Site Code, and Event Release Date Time.

If \code{Biologic_csv}, that implies the data was downloaded from Biologic software in a .csv format.

If \code{raw}, that implies the data was downloaded directly from the reader, in either a .log or .xlsx format. In this case, the largest string containing alphabetic characters in the file name will be assigned as the site code.}
}
\value{
a tibble containing the data downloaded from PTAGIS through a complete capture history query.
}
\description{
The function reads in complete tag histories from either PTAGIS or Biologic that
have been downloaded and coverts the date time values from character
vectors into date-time vectors
}
\examples{
ptagis_file = system.file("extdata", "PRO_Steelhead_2019.csv", package = "PITcleanr")
readCTH(ptagis_file)
}
\author{
Kevin See
}
