% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTagData.R
\name{summariseTagData}
\alias{summariseTagData}
\title{Tag Summary}
\usage{
summariseTagData(proc_list = NULL, saveCSV = F, file_name = NULL)
}
\arguments{
\item{proc_list}{list returned by \code{processCapHist_LGD}.}

\item{saveCSV}{should the resulting dataframe be saved as a csv file? Default value is \code{FALSE}.}

\item{file_name}{if \code{saveCSV} is \code{TRUE}, this is the file path and name, with .csv extension, where the file should be save.}
}
\description{
For each tag, summarises final spawning location as well as important pieces of biological information.
}
\examples{
summariseTagData()
}
\author{
Kevin See
}
