% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLGRNodeNetwork.R
\name{writeLGRNodeNetwork}
\alias{writeLGRNodeNetwork}
\title{LGR Node Network}
\usage{
writeLGRNodeNetwork()
}
\description{
Create data.frame describing all the observation nodes in the
 Lower Granite dam version of DABOM, including how they are related to one
 another. Modified on 1-15-19 to include additonal sites and branches for
 DABOM spawn year 2019
}
\examples{
writeLGRNodeNetwork()
}
\author{
Kevin See
}
